Create dna.py backup
This program takes two files, a CSV file and a string of DNA. It check the CSV file which has a record of individuals and how their DNA patterns in themselves. It compares each persons DNA patterns to the string of DNA to see if they match. If they do it returns the individuals name